renderer_vulkan: Fix incorrect MaxTexelBufferElements return type

+++ src/video_core/renderer_vulkan/vk_instance.h
@@ -229,7 +229,7 @@ public:
     }
 
     /// Returns the maximum supported elements in a texel buffer
-    u32 MaxTexelBufferElements() const {
+    u64 MaxTexelBufferElements() const {
         return properties.limits.maxTexelBufferElements;
     }
